+++ R/RcppExports.R
@@ -454,11 +454,11 @@ forecast_bvharmn <- function(object, step, num_sim) {
 #' @param lag VAR order
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-expand_var <- function(y, lag, include_mean, step, num_iter) {
-    .Call(`_bvhar_expand_var`, y, lag, include_mean, step, num_iter)
+expand_var <- function(y, lag, include_mean, step, y_test) {
+    .Call(`_bvhar_expand_var`, y, lag, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of VHAR based on Expanding Window
@@ -468,11 +468,11 @@ expand_var <- function(y, lag, include_mean, step, num_iter) {
 #' @param y Time series data of which columns indicate the variables
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-expand_vhar <- function(y, include_mean, step, num_iter) {
-    .Call(`_bvhar_expand_vhar`, y, include_mean, step, num_iter)
+expand_vhar <- function(y, include_mean, step, y_test) {
+    .Call(`_bvhar_expand_vhar`, y, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of BVAR based on Expanding Window
@@ -484,11 +484,11 @@ expand_vhar <- function(y, include_mean, step, num_iter) {
 #' @param bayes_spec List, BVAR specification
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-expand_bvar <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
-    .Call(`_bvhar_expand_bvar`, y, lag, bayes_spec, include_mean, step, num_iter)
+expand_bvar <- function(y, lag, bayes_spec, include_mean, step, y_test) {
+    .Call(`_bvhar_expand_bvar`, y, lag, bayes_spec, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of BVAR based on Expanding Window
@@ -500,11 +500,11 @@ expand_bvar <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
 #' @param bayes_spec List, BVAR specification
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-expand_bvarflat <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
-    .Call(`_bvhar_expand_bvarflat`, y, lag, bayes_spec, include_mean, step, num_iter)
+expand_bvarflat <- function(y, lag, bayes_spec, include_mean, step, y_test) {
+    .Call(`_bvhar_expand_bvarflat`, y, lag, bayes_spec, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of BVHAR based on Expanding Window
@@ -515,11 +515,11 @@ expand_bvarflat <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
 #' @param bayes_spec List, BVHAR specification
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-expand_bvhar <- function(y, bayes_spec, include_mean, step, num_iter) {
-    .Call(`_bvhar_expand_bvhar`, y, bayes_spec, include_mean, step, num_iter)
+expand_bvhar <- function(y, bayes_spec, include_mean, step, y_test) {
+    .Call(`_bvhar_expand_bvhar`, y, bayes_spec, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of VAR based on Rolling Window
@@ -530,11 +530,11 @@ expand_bvhar <- function(y, bayes_spec, include_mean, step, num_iter) {
 #' @param lag VAR order
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-roll_var <- function(y, lag, include_mean, step, num_iter) {
-    .Call(`_bvhar_roll_var`, y, lag, include_mean, step, num_iter)
+roll_var <- function(y, lag, include_mean, step, y_test) {
+    .Call(`_bvhar_roll_var`, y, lag, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of VHAR based on Rolling Window
@@ -544,11 +544,11 @@ roll_var <- function(y, lag, include_mean, step, num_iter) {
 #' @param y Time series data of which columns indicate the variables
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-roll_vhar <- function(y, include_mean, step, num_iter) {
-    .Call(`_bvhar_roll_vhar`, y, include_mean, step, num_iter)
+roll_vhar <- function(y, include_mean, step, y_test) {
+    .Call(`_bvhar_roll_vhar`, y, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of BVAR based on Rolling Window
@@ -560,11 +560,11 @@ roll_vhar <- function(y, include_mean, step, num_iter) {
 #' @param bayes_spec List, BVAR specification
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-roll_bvar <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
-    .Call(`_bvhar_roll_bvar`, y, lag, bayes_spec, include_mean, step, num_iter)
+roll_bvar <- function(y, lag, bayes_spec, include_mean, step, y_test) {
+    .Call(`_bvhar_roll_bvar`, y, lag, bayes_spec, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of BVAR based on Rolling Window
@@ -576,11 +576,11 @@ roll_bvar <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
 #' @param bayes_spec List, BVAR specification
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-roll_bvarflat <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
-    .Call(`_bvhar_roll_bvarflat`, y, lag, bayes_spec, include_mean, step, num_iter)
+roll_bvarflat <- function(y, lag, bayes_spec, include_mean, step, y_test) {
+    .Call(`_bvhar_roll_bvarflat`, y, lag, bayes_spec, include_mean, step, y_test)
 }
 
 #' Out-of-Sample Forecasting of BVHAR based on Rolling Window
@@ -591,11 +591,11 @@ roll_bvarflat <- function(y, lag, bayes_spec, include_mean, step, num_iter) {
 #' @param bayes_spec List, BVHAR specification
 #' @param include_mean Add constant term
 #' @param step Integer, Step to forecast
-#' @param num_iter Integer, number of forecasting
+#' @param y_test Evaluation time series data period after `y`
 #' 
 #' @noRd
-roll_bvhar <- function(y, bayes_spec, include_mean, step, num_iter) {
-    .Call(`_bvhar_roll_bvhar`, y, bayes_spec, include_mean, step, num_iter)
+roll_bvhar <- function(y, bayes_spec, include_mean, step, y_test) {
+    .Call(`_bvhar_roll_bvhar`, y, bayes_spec, include_mean, step, y_test)
 }
 
 #' Forecasting Vector Autoregression