+++ src/forecast-expand.cpp
@@ -11,7 +11,7 @@
 //' @param lag VAR order
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -19,10 +19,12 @@ Eigen::MatrixXd expand_var(Eigen::MatrixXd y,
                            int lag, 
                            bool include_mean, 
                            int step,
-                           int num_iter) {
+                           Eigen::MatrixXd y_test) {
   Rcpp::Function fit("var_lm");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd expand_mat(window + num_iter, dim); // train + h-step forecast points
   expand_mat.block(0, 0, window, dim) = y;
   Rcpp::List var_mod = fit(y, lag, include_mean);
@@ -30,7 +32,7 @@ Eigen::MatrixXd expand_var(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    expand_mat.row(window + i) = res.row(i - 1); // expanding window
+    expand_mat.row(window + i - 1) = y_test.row(i - 1); // expanding window
     var_mod = fit(
       expand_mat.block(0, 0, window + i, dim),
       lag,
@@ -49,17 +51,19 @@ Eigen::MatrixXd expand_var(Eigen::MatrixXd y,
 //' @param y Time series data of which columns indicate the variables
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
 Eigen::MatrixXd expand_vhar(Eigen::MatrixXd y, 
                             bool include_mean, 
                             int step,
-                            int num_iter) {
+                            Eigen::MatrixXd y_test) {
   Rcpp::Function fit("vhar_lm");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd expand_mat(window + num_iter, dim); // train + h-step forecast points
   expand_mat.block(0, 0, window, dim) = y;
   Rcpp::List vhar_mod = fit(y, include_mean);
@@ -67,7 +71,7 @@ Eigen::MatrixXd expand_vhar(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    expand_mat.row(window + i) = res.row(i - 1); // expanding window
+    expand_mat.row(window + i - 1) = y_test.row(i - 1); // expanding window
     vhar_mod = fit(
       expand_mat.block(0, 0, window + i, dim),
       include_mean
@@ -87,7 +91,7 @@ Eigen::MatrixXd expand_vhar(Eigen::MatrixXd y,
 //' @param bayes_spec List, BVAR specification
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -96,13 +100,15 @@ Eigen::MatrixXd expand_bvar(Eigen::MatrixXd y,
                             Rcpp::List bayes_spec,
                             bool include_mean, 
                             int step,
-                            int num_iter) {
+                            Eigen::MatrixXd y_test) {
   if (!bayes_spec.inherits("bvharspec")) {
     Rcpp::stop("'object' must be bvharspec object.");
   }
   Rcpp::Function fit("bvar_minnesota");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd expand_mat(window + num_iter, dim); // train + h-step forecast points
   expand_mat.block(0, 0, window, dim) = y;
   Rcpp::List bvar_mod = fit(y, lag, bayes_spec, include_mean);
@@ -111,7 +117,7 @@ Eigen::MatrixXd expand_bvar(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    expand_mat.row(window + i) = res.row(i - 1); // expanding window
+    expand_mat.row(window + i - 1) = y_test.row(i - 1); // expanding window
     bvar_mod = fit(
       expand_mat.block(0, 0, window + i, dim),
       lag, 
@@ -134,7 +140,7 @@ Eigen::MatrixXd expand_bvar(Eigen::MatrixXd y,
 //' @param bayes_spec List, BVAR specification
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -143,13 +149,15 @@ Eigen::MatrixXd expand_bvarflat(Eigen::MatrixXd y,
                                 Rcpp::List bayes_spec,
                                 bool include_mean, 
                                 int step,
-                                int num_iter) {
+                                Eigen::MatrixXd y_test) {
   if (!bayes_spec.inherits("bvharspec")) {
     Rcpp::stop("'object' must be bvharspec object.");
   }
   Rcpp::Function fit("bvar_flat");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd expand_mat(window + num_iter, dim); // train + h-step forecast points
   expand_mat.block(0, 0, window, dim) = y;
   Rcpp::List bvar_mod = fit(y, lag, bayes_spec, include_mean);
@@ -158,7 +166,7 @@ Eigen::MatrixXd expand_bvarflat(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    expand_mat.row(window + i) = res.row(i - 1); // expanding window
+    expand_mat.row(window + i - 1) = y_test.row(i - 1); // expanding window
     bvar_mod = fit(
       expand_mat.block(0, 0, window + i, dim),
       lag, 
@@ -180,7 +188,7 @@ Eigen::MatrixXd expand_bvarflat(Eigen::MatrixXd y,
 //' @param bayes_spec List, BVHAR specification
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -188,13 +196,15 @@ Eigen::MatrixXd expand_bvhar(Eigen::MatrixXd y,
                              Rcpp::List bayes_spec,
                              bool include_mean, 
                              int step,
-                             int num_iter) {
+                             Eigen::MatrixXd y_test) {
   if (!bayes_spec.inherits("bvharspec")) {
     Rcpp::stop("'object' must be bvharspec object.");
   }
   Rcpp::Function fit("bvhar_minnesota");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd expand_mat(window + num_iter, dim); // train + h-step forecast points
   expand_mat.block(0, 0, window, dim) = y;
   Rcpp::List bvhar_mod = fit(y, bayes_spec, include_mean);
@@ -203,7 +213,7 @@ Eigen::MatrixXd expand_bvhar(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    expand_mat.row(window + i) = res.row(i - 1); // expanding window
+    expand_mat.row(window + i - 1) = y_test.row(i - 1); // expanding window
     bvhar_mod = fit(
       expand_mat.block(0, 0, window + i, dim), 
       bayes_spec,