+++ src/forecast-roll.cpp
@@ -11,7 +11,7 @@
 //' @param lag VAR order
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -19,18 +19,20 @@ Eigen::MatrixXd roll_var(Eigen::MatrixXd y,
                          int lag, 
                          bool include_mean, 
                          int step,
-                         int num_iter) {
+                         Eigen::MatrixXd y_test) {
   Rcpp::Function fit("var_lm");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd roll_mat = y; // same size as y
   Rcpp::List var_mod = fit(roll_mat, lag, include_mean);
   Eigen::MatrixXd y_pred = forecast_var(var_mod, step); // step x m
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim);
-    roll_mat.row(window - 1) = res.row(i - 1);
+    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim); // rolling windows
+    roll_mat.row(window - 1) = y_test.row(i - 1); // rolling windows: move one period
     var_mod = fit(roll_mat, lag, include_mean);
     y_pred = forecast_var(var_mod, step);
     res.row(i) = y_pred.row(step - 1);
@@ -45,25 +47,27 @@ Eigen::MatrixXd roll_var(Eigen::MatrixXd y,
 //' @param y Time series data of which columns indicate the variables
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
 Eigen::MatrixXd roll_vhar(Eigen::MatrixXd y, 
                           bool include_mean, 
                           int step,
-                          int num_iter) {
+                          Eigen::MatrixXd y_test) {
   Rcpp::Function fit("vhar_lm");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd roll_mat = y; // same size as y
   Rcpp::List vhar_mod = fit(roll_mat, include_mean);
   Eigen::MatrixXd y_pred = forecast_vhar(vhar_mod, step); // step x m
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim);
-    roll_mat.row(window - 1) = res.row(i - 1);
+    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim); // rolling windows
+    roll_mat.row(window - 1) = y_test.row(i - 1); // rolling windows
     vhar_mod = fit(roll_mat, include_mean);
     y_pred = forecast_vhar(vhar_mod, step);
     res.row(i) = y_pred.row(step - 1);
@@ -80,7 +84,7 @@ Eigen::MatrixXd roll_vhar(Eigen::MatrixXd y,
 //' @param bayes_spec List, BVAR specification
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -89,13 +93,15 @@ Eigen::MatrixXd roll_bvar(Eigen::MatrixXd y,
                           Rcpp::List bayes_spec,
                           bool include_mean, 
                           int step,
-                          int num_iter) {
+                          Eigen::MatrixXd y_test) {
   if (!bayes_spec.inherits("bvharspec")) {
     Rcpp::stop("'object' must be bvharspec object.");
   }
   Rcpp::Function fit("bvar_minnesota");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd roll_mat = y; // same size as y
   Rcpp::List bvar_mod = fit(roll_mat, lag, bayes_spec, include_mean);
   Rcpp::List bvar_pred = forecast_bvar(bvar_mod, step, 1);
@@ -103,8 +109,8 @@ Eigen::MatrixXd roll_bvar(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim);
-    roll_mat.row(window - 1) = res.row(i - 1);
+    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim); // rolling windows
+    roll_mat.row(window - 1) = y_test.row(i - 1); // rolling windows
     bvar_mod = fit(roll_mat, lag, bayes_spec, include_mean);
     bvar_pred = forecast_bvar(bvar_mod, step, 1);
     y_pred = bvar_pred["posterior_mean"];
@@ -122,7 +128,7 @@ Eigen::MatrixXd roll_bvar(Eigen::MatrixXd y,
 //' @param bayes_spec List, BVAR specification
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -131,13 +137,15 @@ Eigen::MatrixXd roll_bvarflat(Eigen::MatrixXd y,
                               Rcpp::List bayes_spec,
                               bool include_mean, 
                               int step,
-                              int num_iter) {
+                              Eigen::MatrixXd y_test) {
   if (!bayes_spec.inherits("bvharspec")) {
     Rcpp::stop("'object' must be bvharspec object.");
   }
   Rcpp::Function fit("bvar_flat");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd roll_mat = y; // same size as y
   Rcpp::List bvar_mod = fit(roll_mat, lag, bayes_spec, include_mean);
   Rcpp::List bvar_pred = forecast_bvar(bvar_mod, step, 1);
@@ -145,8 +153,8 @@ Eigen::MatrixXd roll_bvarflat(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim);
-    roll_mat.row(window - 1) = res.row(i - 1);
+    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim); // rolling windows
+    roll_mat.row(window - 1) = y_test.row(i - 1); // rolling windows
     bvar_mod = fit(roll_mat, lag, bayes_spec, include_mean);
     bvar_pred = forecast_bvar(bvar_mod, step, 1);
     y_pred = bvar_pred["posterior_mean"];
@@ -163,7 +171,7 @@ Eigen::MatrixXd roll_bvarflat(Eigen::MatrixXd y,
 //' @param bayes_spec List, BVHAR specification
 //' @param include_mean Add constant term
 //' @param step Integer, Step to forecast
-//' @param num_iter Integer, number of forecasting
+//' @param y_test Evaluation time series data period after `y`
 //' 
 //' @noRd
 // [[Rcpp::export]]
@@ -171,13 +179,15 @@ Eigen::MatrixXd roll_bvhar(Eigen::MatrixXd y,
                            Rcpp::List bayes_spec,
                            bool include_mean, 
                            int step,
-                           int num_iter) {
+                           Eigen::MatrixXd y_test) {
   if (!bayes_spec.inherits("bvharspec")) {
     Rcpp::stop("'object' must be bvharspec object.");
   }
   Rcpp::Function fit("bvhar_minnesota");
   int window = y.rows();
   int dim = y.cols();
+  int num_test = y_test.rows();
+  int num_iter = num_test - step + 1; // longest forecast horizon
   Eigen::MatrixXd roll_mat = y; // same size as y
   Rcpp::List bvhar_mod = fit(roll_mat, bayes_spec, include_mean);
   Rcpp::List bvhar_pred = forecast_bvharmn(bvhar_mod, step, 1);
@@ -185,8 +195,8 @@ Eigen::MatrixXd roll_bvhar(Eigen::MatrixXd y,
   Eigen::MatrixXd res(num_iter, dim);
   res.row(0) = y_pred.row(step - 1); // only need the last one (e.g. step = h => h-th row)
   for (int i = 1; i < num_iter; i++) {
-    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim);
-    roll_mat.row(window - 1) = res.row(i - 1);
+    roll_mat.block(0, 0, window - 1, dim) = roll_mat.block(1, 0, window - 1, dim); // rolling windows
+    roll_mat.row(window - 1) = y_test.row(i - 1); // rolling windows
     bvhar_mod = fit(roll_mat, bayes_spec, include_mean);
     bvhar_pred = forecast_bvharmn(bvhar_mod, step, 1);
     y_pred = bvhar_pred["posterior_mean"];